requirements.txt
torch>=1.10.0
torchvision>=0.11.1
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
h5py>=3.1.0
biopython>=1.79
umap-learn>=0.5.1
tqdm>=4.62.0
pyyaml>=6.0
joblib>=1.1.0
google-generativeai>=0.1.0; python_version>="3.9"